backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
spacy==3.6.1
scikit-learn==1.3.1
PyPDF2==3.0.1
python-docx==0.8.11
werkzeug==2.3.7
